+++ CHANGELOG.md
@@ -36,7 +36,8 @@ Advantages:
 - [frontend]: Added Log length: 100, 200, 500, 1000.
 - [frontend]: Added Log search with case-insensitive regex: use /text/ in the search field.
 - [frontend]: Differentiated update icon color: primary color for latest and secondary color for dev.
-- [MbfTable]: Added ConditionalTooltip to show clipped column cells.
+- [frontend]: Added `Reset the frontend UI` in the Reset menu. It will reset the frontend local storage to the default.
+- [MbfTable]: Added ConditionalTooltip to show clipped column cells. Used in the Cluster table.
 
 ### Changed